- moved loadGlobalStrings() to globals.js - call loadGlobalStrings() early on to init values
